implemented method for calculating total points

+++ SoftwareDesignProject/SoftwareDesignProject/Data/Models/Student.cs
@@ -12,7 +12,7 @@ namespace SoftwareDesignProject.Data.Models
         public string EmriIPrindit { get; set; }
         public string Mbiemri { get; set; }
         public string Qyteti { get; set; }
-        public float NotaMesatare { get; set; }
+        public double NotaMesatare { get; set; }
         public int NumriKontaktues { get; set; }
         public string Email { get; set; }
         public char Gjinia { get; set; }
